Fix progress logging for remote projections
Prior to this commit, an adaptive scheme was used, that increased the
task volume with each processed batch.
That scheme resulted in progress reaching 100% after one batch, 50%
after two batches, 33% after three, and so on.
Generally, the progress would be 1/b where b is the batch ordinal.
For a large amount of batches, progress would reach 0%.

Instead, we now use the Iterative.OPEN task mode.
We do not know the total amount of incoming rows.
Instead, we log progress after completing each batch.