Don't count relationships by there highest id
On block format stores, the relationship id has a different meaning and
is not representative of a "highest possible relationship count".

Since we use that value only for logging the import progress, it is ok
if we end up being slightly inaccurate, so we can use the number of
existing relationship as the upper bound.