Fallback to ArrayUtil.oversize if we overvflow
The default way is to always grow the byte[] to the
next power of two of the requested length. This can
overflow even if there might still be enough space.
We now fall back to Lucene's oversize implementation
in case of an overflow and try if we can make it fit.