work around release ops bug
release ops provide us with docker image that runs neo4j with version 5.21.0 instead of 5.20.0.
they did promise that this is indeed the version to be released, and that this is indeed 5.20.0 and not 5.21.0, so what we're gonna do for now to not block the releae, we're treat it as if they sent the right version number of 5.20.0.