Consolidate down to a single Neo4j version
OpenGDS was the use case for keeping one stable and one next.
But the next requires breaking with stable, and next is not possible to
use with openGDS. So we're not Public repo syncing anymore.
Then we may as well not have a red openGDS build for the whole month,
but we build also openGDS with next. Then we sync at end of month.