we need to register exporter builder components.
There was a hope that we would only need one (global) extension. It is clear now that we need two: one global, one for per-database things.
So here we simplify and document that state of affairs, so that it is clear for posterity what is going on and why things are safe.