have progress tracking module define it's own logger
the logger might actually better be described as an "eventer", because the events it emits are in turn sent to a textual log, but also to the in-memory task tracker thingie. something for someone to clean up one day when we revolutionise the progress tracker module.

surround usages of the new logger with an adapter so we can work in stages. adapter can stay for spec framework since there are no use cases for isolated/ first class progress locking there - the use case that we are chasing here is the ability to turn off progress logging in a nice way.