We don't have neo4j
We are usually working around this by having a copy of the actual config
which overrides the fields that are using problematic converters to use
something else.
However, with the refactor around configs, the generated builders don't
really work anymore and I wanted to use a config builder to build the
individual dijkstra configs.
This is a bit of a hack, but it works, soooo 😎